[ISSUE #3671] move some tools class into common package (#3672)
